Comprobar que ninguna variable esté vacía para evitar errores inesperados

+++ scripts/bin/setup-general.sh
@@ -10,41 +10,41 @@ export LC_ALL=C
 # Execute ERB files
 recursiveExecuteErbs() {
 	for path in "${1:?}"/*; do
-		if [ -d "${path}" ] && [ ! -L "${path}" ]; then
-			recursiveExecuteErbs "${path}"
-		elif [ "${path}" != "${path%.erb}" ]; then
-			logInfo "Executing ERB file: ${path}"
-			dirname=${path%/*}; basename=$(basename "${path}" .erb)
-			output=${dirname}/${basename}
-			erb -T - -- "${output}.erb" > "${output}"
-			chmod --reference="${output}.erb" -- "${output}"
+		if [ -d "${path:?}" ] && [ ! -L "${path:?}" ]; then
+			recursiveExecuteErbs "${path:?}"
+		elif [ "${path:?}" != "${path%.erb}" ]; then
+			logInfo "Executing ERB file: ${path:?}"
+			dirname=${path%/*}; basename=$(basename "${path:?}" .erb)
+			output=${dirname:?}/${basename:?}
+			erb -T - -- "${output:?}.erb" > "${output:?}"
+			chmod --reference="${output:?}.erb" -- "${output:?}"
 		fi
 	done
 }
-recursiveExecuteErbs "${BISERVER_HOME}"
+recursiveExecuteErbs "${BISERVER_HOME:?}"
 
 # Compress directories ending in .zip, .pfm or .pgus
 recursiveZipDirs() {
 	for path in "${1:?}"/*; do
-		if [ -d "${path}" ] && [ ! -L "${path}" ]; then
+		if [ -d "${path:?}" ] && [ ! -L "${path:?}" ]; then
 			# This method must be called in a subshell to avoid
 			# overwriting variables in the current scope
-			(recursiveZipDirs "${path}")
+			(recursiveZipDirs "${path:?}")
 			if
-				[ "${path}" != "${path%.zip}" ] ||
+				[ "${path:?}" != "${path%.zip}" ] ||
 				# Pentaho File Metadata plugin
-				[ "${path}" != "${path%.pfm}" ] ||
+				[ "${path:?}" != "${path%.pfm}" ] ||
 				# Pentaho Global User Settings plugin
-				[ "${path}" != "${path%.pgus}" ]
+				[ "${path:?}" != "${path%.pgus}" ]
 			then
-				logInfo "Compressing directory: ${path}"
+				logInfo "Compressing directory: ${path:?}"
 				dirname=${path%/*}; basename=${path##*/}
-				output=${dirname}/${basename}; tmpOutput=$(mktemp -u)
-				cd -- "${output}" || exit; zip -qmyr "${tmpOutput}" ./
-				cd -- "${OLDPWD}" || exit; rmdir -- "${output}"
-				[ -f "${tmpOutput}" ] && mv -- "${tmpOutput}" "${output}"
+				output=${dirname:?}/${basename:?}; tmpOutput=$(mktemp -u)
+				cd -- "${output:?}" || exit; zip -qmyr "${tmpOutput:?}" ./
+				cd -- "${OLDPWD:?}" || exit; rmdir -- "${output:?}"
+				[ -f "${tmpOutput:?}" ] && mv -- "${tmpOutput:?}" "${output:?}"
 			fi
 		fi
 	done
 }
-recursiveZipDirs "${BISERVER_HOME}"
+recursiveZipDirs "${BISERVER_HOME:?}"
